airflow/requirements.txt
requests==2.31.0
pandas==2.2.2
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
tenacity==8.2.3
ratelimit==2.2.1
python-dotenv==1.0.1
